syncslides: Port account and blessing initialization.
V23.java is heavily inspired by the old V23Manager.java.  I stripped out
anything related to mounttables, and I simplified some of the code.

SignInActivity and related files are a copy/paste from the other project.
This activity requires a few more permissions and a min sdk of 22.

Also, renamed the app to DevSyncSlides so I can retain the demo-ready
version while I do development on this version.

Change-Id: Ia7a468a76c7df492e8f3d36f8528c48f8299c159