syncslides: Move DeckChooserActivity.
Move DeckChooserActivity and related fragments to the new project.
The main functionality of the activity is commented out, but the
navigation drawer works.

I moved all of the resources over all at once rather than trying to
do it selectively.  I don't expect to change any of these resources
as part of refactoring the code, so they will all be used as is.

Change-Id: Ia67871e89636af840c10364545ffa1836ff6198d