syncslides: Updating to the latest versions of Flutter, Mojo and v23Discovery and fixing breaking changes.
-Adding some logging.
-Using asset bundl so images load when phone is
not connected Fixes: https://github.com/vanadium/syncslides/issues/1

Please note that Flutter no longer publishes on Pub and
therefore we all need to have a clone of Flutter at a
specific place (until we add a Flutter profile)

I have added two files that specify the commit # for
flutter and mojo that I have tested the app in. This is
temporary and will go away when we have Mojo and Flutter
profiles.

Change-Id: If658a9cebb3fb649686dcfb5b88dfc164b74c6c9