syncslides: Session and Presentation interfaces.
The Session holds the deckId and presentationId, paramters that
we stored in the Bundle in the old version of Syncslides.  The
plan is to add more functionality to the Session such as the
ability to watch the local slide number for changes.

The Presentation holds all of the state for the presentation that
the user is currently viewing.  The plan is for SyncbaseDB to cache
the most recently used Presentation so that when the user rotates
their phone, we don't need to refetch all of the slide data from
the database-- it'll be right there in the Presentation.

This change adds the first of many methods to Presentation: getSlides().
I plan to use this method in the slide list view of the app.
I didn't include all of that UI code in this change because it
was getting too big.

Change-Id: Iab053dea8798450b9107338bd32033b584726120