syncslides: Marshmallow requires dynamic permission requests.
Android 6.0 (Marshmallow) requires that the app ask at runtime
for permission to access the contacts.  This change switches
the code to do that.  Rather than trying to support devices
older than 6.0, the app now requires 6.0.

Changed the build.gradle applicationId to io.v.syncslides23
because one of my devices refused to let me upload an apk
with the old name.  Ugh.

Change-Id: I96bf5f47198a078baf072084747139ef6381bd8d